requirements.txt
backports.functools-lru-cache==1.6.1
blinker==1.4
click==7.1.2
configobj==5.0.6
cycler==0.10.0
decorator==4.4.2
Flask==1.1.2
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gunicorn==19.10.0

itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
networkx==2.2
numpy==1.16.6
passlib==1.7.4
Pillow==6.2.2
psutil==5.4.2
pycairo==1.16.2
pycrypto==2.6.1

psycopg2==2.8.6
boto3==1.16.17
scikit-image==0.17.2
python-xlib==0.20
PyWavelets==1.0.3
pyxdg==0.25
scipy==1.2.3
setproctitle==1.1.10
six==1.15.0
SQLAlchemy==1.3.20
Werkzeug==1.0.1
WTForms==2.3.3